corrected windows to be doubles.

+++ probability_model.hpp
@@ -113,7 +113,7 @@ class probability_model{
   bool m_owner_of_seasonal_scale;
   bool m_owner_of_time_scale;
   unsigned int m_num_windows;
-  unsigned int* m_windows;
+  double* m_windows;
   double** m_windowed_lhd_contributions;
   double* m_window_mixture_probs;
   double m_window_mixture_probs_sum;
